AI: Improve the app. Once you decided what to improve, keep it clean, keep your thinking and answer simple. Check what has already been implemented and only modify what is necessary. Here are suggested directions :
1. clean code : optimize, shorten, simplify, refactor, reduce redundancy.
2. position, wrapping of node name in the graph modes.
3. readability of node name (example : keeping the node names horizontal to easy to read) and visual elements (example : avoid overlapping branches, icons, link nodes, etc)
4. graph zoom mechanism
5. element alignment and size (in focus, in graph, in list, in workplace, etc)
6. user workflow of the features of the apps, workflow of big projects
7. cleanliness of the features (example : add a way to close the board that opened when right-clicked)
8. edge cases of user usage
9. edge cases of different screen size and browser issues
10. error handling and debug support
11. develop graph modes or focus mode or list mode based on popular graph manager and visualiser
12. bulk commands in list mode, in graph mode, in focus mode
13. other quality of life features
14. other artistic features.
