+++ js/components/GraphViewComponent.js
@@ -2,146 +2,7 @@ import React, { useEffect, useRef, useMemo, useCallback, useState } from 'react'
 import { linkRadial, select, linkVertical, linkHorizontal } from 'd3';
 import { useD3Tree } from '../hooks/useD3Tree.js';
 import { NODE_IMPORTANCE_RUNES } from '../constants.js';
-
-// A robust text wrapping function for SVG <text> elements.
-function wrap(textSelection, width, maxLines = 3) {
-    const lineHeight = 1.2; // ems
-
-    textSelection.each(function() {
-        const text = select(this);
-        const words = text.text().split(/\s+/).reverse();
-        text.text(null); // Clear original text content
-
-        let tspan = text.append('tspan').attr('x', 0);
-        let line = [];
-        let lineNumber = 0;
-        let word;
-
-        // Core wrapping loop
-        while ((word = words.pop())) {
-            line.push(word);
-            tspan.text(line.join(' '));
-
-            // If line is too long and has more than one word, wrap it
-            if (tspan.node().getComputedTextLength() > width && line.length > 1) {
-                line.pop(); // remove the word that made it too long
-                tspan.text(line.join(' '));
-
-                if (lineNumber + 1 < maxLines) {
-                    lineNumber++;
-                    line = [word]; // start new line with the popped word
-                    tspan = text.append('tspan').attr('x', 0).attr('dy', `${lineHeight}em`).text(word);
-                } else {
-                    // We are on the last allowed line and it's full. Add ellipsis and stop.
-                    tspan.text(tspan.text() + '…');
-                    words.length = 0; // empty the words array to exit loop
-                    break;
-                }
-            }
-        }
-        
-        // Post-loop cleanup for ellipsis if we exited due to maxLines
-        if (words.length > 0 && !tspan.text().endsWith('…')) {
-            tspan.text(tspan.text() + '…');
-        }
-
-        // Edge case: Handle a single word that is longer than the allowed width
-        if (text.selectAll('tspan').size() === 1 && tspan.node().getComputedTextLength() > width) {
-            let currentText = tspan.text();
-            while(tspan.node().getComputedTextLength() > width && currentText.length > 3) {
-                currentText = currentText.slice(0, -1);
-                tspan.text(currentText + '…');
-            }
-        }
-
-        // Adjust vertical alignment for multi-line text based on dominant-baseline
-        const numLines = text.selectAll('tspan').size();
-        if (numLines > 1) {
-            const dominantBaseline = text.attr('dominant-baseline');
-            const firstTspan = text.select('tspan');
-            const totalExtraHeightEms = (numLines - 1) * lineHeight;
-            
-            if (dominantBaseline === 'middle') {
-                // Shift the entire text block up by half its height
-                firstTspan.attr('dy', `-${totalExtraHeightEms / 2}em`);
-            } else if (dominantBaseline === 'baseline' || dominantBaseline === 'auto' || dominantBaseline === 'alphabetic' || dominantBaseline === 'text-after-edge') {
-                // For baselines at the bottom of the text, shift the whole block up.
-                firstTspan.attr('dy', `-${totalExtraHeightEms}em`);
-            }
-            // 'hanging' and 'text-before-edge' are fine by default, as text flows down from the top.
-        }
-    });
-}
-
-// A robust text wrapping function for SVG <text> elements.
-function wrap(textSelection, width, maxLines = 3) {
-    const lineHeight = 1.2; // ems
-
-    textSelection.each(function() {
-        const text = select(this);
-        const words = text.text().split(/\s+/).reverse();
-        text.text(null); // Clear original text content
-
-        let tspan = text.append('tspan').attr('x', 0);
-        let line = [];
-        let lineNumber = 0;
-        let word;
-
-        // Core wrapping loop
-        while ((word = words.pop())) {
-            line.push(word);
-            tspan.text(line.join(' '));
-
-            // If line is too long and has more than one word, wrap it
-            if (tspan.node().getComputedTextLength() > width && line.length > 1) {
-                line.pop(); // remove the word that made it too long
-                tspan.text(line.join(' '));
-
-                if (lineNumber + 1 < maxLines) {
-                    lineNumber++;
-                    line = [word]; // start new line with the popped word
-                    tspan = text.append('tspan').attr('x', 0).attr('dy', `${lineHeight}em`).text(word);
-                } else {
-                    // We are on the last allowed line and it's full. Add ellipsis and stop.
-                    tspan.text(tspan.text() + '…');
-                    words.length = 0; // empty the words array to exit loop
-                    break;
-                }
-            }
-        }
-        
-        // Post-loop cleanup for ellipsis if we exited due to maxLines
-        if (words.length > 0 && !tspan.text().endsWith('…')) {
-            tspan.text(tspan.text() + '…');
-        }
-
-        // Edge case: Handle a single word that is longer than the allowed width
-        if (text.selectAll('tspan').size() === 1 && tspan.node().getComputedTextLength() > width) {
-            let currentText = tspan.text();
-            while(tspan.node().getComputedTextLength() > width && currentText.length > 3) {
-                currentText = currentText.slice(0, -1);
-                tspan.text(currentText + '…');
-            }
-        }
-
-        // Adjust vertical alignment for multi-line text based on dominant-baseline
-        const numLines = text.selectAll('tspan').size();
-        if (numLines > 1) {
-            const dominantBaseline = text.attr('dominant-baseline');
-            const firstTspan = text.select('tspan');
-            const totalExtraHeightEms = (numLines - 1) * lineHeight;
-            
-            if (dominantBaseline === 'middle') {
-                // Shift the entire text block up by half its height
-                firstTspan.attr('dy', `-${totalExtraHeightEms / 2}em`);
-            } else if (dominantBaseline === 'baseline' || dominantBaseline === 'auto' || dominantBaseline === 'alphabetic' || dominantBaseline === 'text-after-edge') {
-                // For baselines at the bottom of the text, shift the whole block up.
-                firstTspan.attr('dy', `-${totalExtraHeightEms}em`);
-            }
-            // 'hanging' and 'text-before-edge' are fine by default, as text flows down from the top.
-        }
-    });
-}
+import { wrapSvgText } from '../utils.js';
 
 
 const getNodeRadius = (node) => {
@@ -489,7 +350,7 @@ const GraphViewComponent = ({
             const radius = getNodeRadius(d);
             const maxTextWidth = radius * 5; // Proportional to node size
             const maxLines = 3;
-            wrap(select(this), maxTextWidth, maxLines);
+            wrapSvgText(select(this), maxTextWidth, maxLines);
         }
       });
