+++ js/utils.js
@@ -1,4 +1,5 @@
 import { v4 as uuidv4 } from 'uuid';
+import { select } from 'd3';
 
 export const generateUUID = () => uuidv4();
 
@@ -273,6 +274,81 @@ export const downloadObjectAsJson = (exportObj, exportName) => {
   link.click();
   document.body.removeChild(link);
   URL.revokeObjectURL(url);
+};
+
+/**
+ * A robust text wrapping function for SVG <text> elements using D3.
+ * @param {d3.Selection} textSelection - The D3 selection of <text> elements to wrap.
+ * @param {number} width - The maximum width in pixels for the text.
+ * @param {number} [maxLines=3] - The maximum number of lines to wrap to.
+ */
+export const wrapSvgText = (textSelection, width, maxLines = 3) => {
+    const lineHeight = 1.2; // ems
+
+    textSelection.each(function() {
+        const text = select(this);
+        const words = text.text().split(/\s+/).reverse();
+        text.text(null); // Clear original text content
+
+        let tspan = text.append('tspan').attr('x', 0);
+        let line = [];
+        let lineNumber = 0;
+        let word;
+
+        // Core wrapping loop
+        while ((word = words.pop())) {
+            line.push(word);
+            tspan.text(line.join(' '));
+
+            // If line is too long and has more than one word, wrap it
+            if (tspan.node().getComputedTextLength() > width && line.length > 1) {
+                line.pop(); // remove the word that made it too long
+                tspan.text(line.join(' '));
+
+                if (lineNumber + 1 < maxLines) {
+                    lineNumber++;
+                    line = [word]; // start new line with the popped word
+                    tspan = text.append('tspan').attr('x', 0).attr('dy', `${lineHeight}em`).text(word);
+                } else {
+                    // We are on the last allowed line and it's full. Add ellipsis and stop.
+                    tspan.text(tspan.text() + '…');
+                    words.length = 0; // empty the words array to exit loop
+                    break;
+                }
+            }
+        }
+        
+        // Post-loop cleanup for ellipsis if we exited due to maxLines
+        if (words.length > 0 && !tspan.text().endsWith('…')) {
+            tspan.text(tspan.text() + '…');
+        }
+
+        // Edge case: Handle a single word that is longer than the allowed width
+        if (text.selectAll('tspan').size() === 1 && tspan.node().getComputedTextLength() > width) {
+            let currentText = tspan.text();
+            while(tspan.node().getComputedTextLength() > width && currentText.length > 3) {
+                currentText = currentText.slice(0, -1);
+                tspan.text(currentText + '…');
+            }
+        }
+
+        // Adjust vertical alignment for multi-line text based on dominant-baseline
+        const numLines = text.selectAll('tspan').size();
+        if (numLines > 1) {
+            const dominantBaseline = text.attr('dominant-baseline');
+            const firstTspan = text.select('tspan');
+            const totalExtraHeightEms = (numLines - 1) * lineHeight;
+            
+            if (dominantBaseline === 'middle') {
+                // Shift the entire text block up by half its height
+                firstTspan.attr('dy', `-${totalExtraHeightEms / 2}em`);
+            } else if (dominantBaseline === 'baseline' || dominantBaseline === 'auto' || dominantBaseline === 'alphabetic' || dominantBaseline === 'text-after-edge') {
+                // For baselines at the bottom of the text, shift the whole block up.
+                firstTspan.attr('dy', `-${totalExtraHeightEms}em`);
+            }
+            // 'hanging' and 'text-before-edge' are fine by default, as text flows down from the top.
+        }
+    });
 };
 
 export const compareAndAnnotateTree = (originalTree, modifiedTree) => {
